refactor SetUtility.kt into /util
